pytest: Separate HL7 tests
