+++ server/camcops_server/cc_modules/cc_hl7.py
@@ -73,7 +73,6 @@ from pendulum import Date, DateTime as Pendulum
 
 from camcops_server.cc_modules.cc_constants import DateFormat, FileType
 from camcops_server.cc_modules.cc_simpleobjects import HL7PatientIdentifier
-from camcops_server.cc_modules.cc_unittest import DemoDatabaseTestCase
 
 if TYPE_CHECKING:
     from camcops_server.cc_modules.cc_request import CamcopsRequest
@@ -866,58 +865,3 @@ class MLLPTimeoutClient(object):
             return True, ack_msg
         except socket.timeout:
             return False, None
-
-
-# =============================================================================
-# Unit tests
-# =============================================================================
-
-class HL7CoreTests(DemoDatabaseTestCase):
-    """
-    Unit tests.
-    """
-    def test_hl7core_func(self) -> None:
-        from camcops_server.cc_modules.cc_simpleobjects import TaskExportOptions  # noqa
-        from camcops_server.tasks.phq9 import Phq9
-
-        self.announce("test_hl7core_func")
-
-        pitlist = [
-            HL7PatientIdentifier(pid="1", id_type="TT",
-                                 assigning_authority="AA")
-        ]
-        # noinspection PyTypeChecker
-        dob = Date.today()  # type: Date
-        now = Pendulum.now()
-        task = self.dbsession.query(Phq9).first()
-        assert task, "Missing Phq9 in demo database!"
-
-        self.assertIsInstance(get_mod11_checkdigit("12345"), str)
-        self.assertIsInstance(get_mod11_checkdigit("badnumber"), str)
-        self.assertIsInstance(get_mod11_checkdigit("None"), str)
-        self.assertIsInstance(make_msh_segment(now, "control_id"), hl7.Segment)
-        self.assertIsInstance(make_pid_segment(
-            forename="fname",
-            surname="sname",
-            dob=dob,
-            sex="M",
-            address="Somewhere",
-            patient_id_list=pitlist
-        ), hl7.Segment)
-        self.assertIsInstance(make_obr_segment(task), hl7.Segment)
-        for task_format in [FileType.PDF, FileType.HTML, FileType.XML]:
-            for comments in [True, False]:
-                export_options = TaskExportOptions(
-                    xml_include_comments=comments,
-                    xml_with_header_comments=comments,
-                )
-                self.assertIsInstance(make_obx_segment(
-                    req=self.req,
-                    task=task,
-                    task_format=task_format,
-                    observation_identifier="obs_id",
-                    observation_datetime=now,
-                    responsible_observer="responsible_observer",
-                    export_options=export_options,
-                ), hl7.Segment)
-        self.assertIsInstance(escape_hl7_text("blahblah"), str)
